requirements.txt
aiogram==3.17.0
aiosqlite==0.21.0
anthropic==0.45.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
